+++ dubbo-filter-extensions/dubbo-filter-polaris-dubbo2/dubbo-filter-polaris-circuitbreaker-dubbo2/src/main/resources/META-INF/dubbo/org.apache.dubbo.rpc.Filter
@@ -1,0 +1,1 @@
+polaris_circuitbreaker=org.apache.dubbo.filter.dubbo2.CircuitBreakerFilter
