+++ dubbo-filter-extensions/dubbo-filter-polaris-dubbo2/dubbo-filter-polaris-ratelimit-dubbo2/pom.xml
@@ -19,33 +19,12 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>dubbo-filter-extensions</artifactId>
+        <artifactId>dubbo-filter-polaris-dubbo2</artifactId>
         <groupId>org.apache.dubbo.extensions</groupId>
-        <version>${revision}</version>
-        <relativePath>../pom.xml</relativePath>
+        <version>1.0.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>dubbo-filter-polaris-dubbo2</artifactId>
-    <packaging>pom</packaging>
-    <name>dubbo-filter-polaris-dubbo2</name>
-    <version>1.0.0-SNAPSHOT</version>
-    <description>Dubbo2 filter extension for PolarisMesh, support circuitbreaking, ratelimit, metric capabilities.</description>
-    <modules>
-        <module>dubbo-filter-polaris-circuitbreaker-dubbo2</module>
-        <module>dubbo-filter-polaris-ratelimit-dubbo2</module>
-    </modules>
+    <artifactId>dubbo-filter-polaris-ratelimit-dubbo2</artifactId>
 
-    <dependencies>
-        <dependency>
-            <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo</artifactId>
-            <version>2.7.18</version>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>com.tencent.polaris</groupId>
-            <artifactId>polaris-adapter-dubbo</artifactId>
-        </dependency>
-    </dependencies>
 </project>
