+++ dubbo-filter-extensions/dubbo-filter-polaris-dubbo2/dubbo-filter-polaris-ratelimit-dubbo2/src/main/resources/META-INF/dubbo/org.apache.dubbo.rpc.Filter
@@ -1,0 +1,1 @@
+polaris_ratelimit=org.apache.dubbo.filter.dubbo2.RateLimitFilter
