--- dubbo-filter-extensions/dubbo-filter-polaris-dubbo2/src/main/resources/META-INF/dubbo/org.apache.dubbo.rpc.Filter
@@ -1,3 +1,0 @@
-polaris_report=org.apache.dubbo.filter.dubbo2.ReportFilter
-polaris_circuitbreaker=org.apache.dubbo.filter.dubbo2.CircuitBreakerFilter
-polaris_ratelimit=org.apache.dubbo.filter.dubbo2.RateLimitFilter
