--- dubbo-registry-extensions/dubbo-registry-polaris/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.registry.client.ServiceDiscoveryFactory
@@ -1,1 +1,0 @@
-polaris=org.apache.dubbo.registry.polaris.PolarisServiceDiscoveryFactory
