+++ dubbo-registry-extensions/dubbo-registry-polaris/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.Filter
@@ -1,0 +1,1 @@
+polaris_report=org.apache.dubbo.registry.polaris.filter.ReportFilter
